# api_spec.json
{"openapi":"3.1.0","info":{"title":"MockAPI","version":"0.1.0"},"servers":[{"url":"http://127.0.0.1:8000","description":"Mock Server"}],"paths":{"/":{"get":{"summary":"Get User","operationId":"get_user__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}}}}},"/portfolios":{"get":{"summary":"Get Portfolios","operationId":"get_portfolios_portfolios_get","parameters":[{"required":true,"schema":{"type":"string","title":"Currency Code"},"name":"currency_code","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/Portfolio"},"type":"array","title":"Response Get Portfolios Portfolios Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/holdings":{"get":{"summary":"Get Holdings","operationId":"get_holdings_holdings_get","parameters":[{"required":true,"schema":{"type":"integer","title":"Portfolio Id"},"name":"portfolio_id","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/Holding"},"type":"array","title":"Response Get Holdings Holdings Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/performance":{"get":{"summary":"Get Performance","operationId":"get_performance_performance_get","parameters":[{"required":true,"schema":{"type":"integer","title":"Holding Id"},"name":"holding_id","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"number","title":"Response Get Performance Performance Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"Holding":{"properties":{"id":{"type":"integer","title":"Id"},"portfolio":{"$ref":"#/components/schemas/Portfolio"},"date":{"type":"string","format":"date-time","title":"Date"},"instrument":{"$ref":"#/components/schemas/Instrument"},"weight":{"type":"number","title":"Weight"},"performance":{"type":"number","title":"Performance"}},"type":"object","required":["id","portfolio","date","instrument","weight","performance"],"title":"Holding"},"Instrument":{"properties":{"name":{"type":"string","title":"Name"}},"type":"object","required":["name"],"title":"Instrument"},"Portfolio":{"properties":{"id":{"type":"integer","title":"Id"},"code":{"type":"string","title":"Code"},"currency":{"type":"string","title":"Currency"}},"type":"object","required":["id","code","currency"],"title":"Portfolio"},"User":{"properties":{"name":{"type":"string","title":"Name"},"email":{"type":"string","title":"Email"}},"type":"object","required":["name","email"],"title":"User"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}